+++ changelog.md
@@ -5,6 +5,7 @@ vX.X.X
 - Added installation steps for all platforms, closes #184.
 - Fixed bug where 'Open' button on TX/RX tab is enabled when no COM ports can be found, closes #185.
 - Removed manual buffer size limitations on totalNewLineParserOutput, closes #178.
+- Fixed bug where WebView data pane labels where not being setup correctly on start-up, closes #187.
 
 v0.8.5
 ------